refactor: extract SplitGuestOffset into meta/addr.rs
Move the guest-offset address-splitting type out of meta/mod.rs into a
dedicated meta/addr.rs submodule, re-exported from mod.rs. No logic
changes.
